requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.4.0
autogen-core==0.7.5
autogen-ext==0.7.5
azure-ai-inference==1.0.0b9
azure-core==1.38.0
azure-identity==1.25.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.3
flash==1.0.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonargparse==4.42.0
jsonref==1.1.0
kiwisolver==1.4.9
lightning-flash==0.8.2
lightning-utilities==0.15.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.7.0
networkx==3.5
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==1.108.1
opentelemetry-api==1.38.0
packaging==25.0
pandas==2.3.3
pillow==12.1.1
propcache==0.4.1
protobuf==5.29.6
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyDeprecate==0.3.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.4.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.9.0
torchmetrics==0.10.3
tqdm==4.67.1
triton==3.5.0
typeshed_client==2.8.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
yarl==1.22.0
zipp==3.23.0
